requirements.txt
torch
transformers
faiss-cpu
sentence-transformers
pyttsx3
SpeechRecognition
pyaudio
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
selenium
jupyter
textblob
numpy
requests
beautifulsoup4
librosa
soundfile
scipy
python-dotenv
twilio
flask
